+++ src/signal_bus.gd
@@ -1,9 +1,8 @@
 extends Node
 
-signal seedling_grown
 signal flower_bloomed
-signal flower_gone_to_seed
-signal flower_clipped
-signal bee_arrived
 signal cut_flower_decayed
 signal plant_died
+signal flower_pollinated
+signal watering_can_emptied
+signal watering_can_refilled